+++ src/main/java/io/vertx/core/spi/tls/SslProvider.java
@@ -10,11 +10,7 @@
  */
 package io.vertx.core.spi.tls;
 
-import io.vertx.core.buffer.Buffer;
-import io.vertx.core.net.KeyCertOptions;
 import io.vertx.core.net.SSLEngineOptions;
-import io.vertx.core.net.TCPSSLOptions;
-import io.vertx.core.net.TrustOptions;
 
 import java.util.List;
 import java.util.Set;
@@ -27,10 +23,6 @@ import java.util.Set;
 public interface SslProvider {
 
   SslContextFactory contextFactory(SSLEngineOptions options,
-                                   KeyCertOptions keyCertOptions,
-                                   TrustOptions trustOptions,
-                                   List<String> crltPaths,
-                                   List<Buffer> crlValues,
                                    Set<String> enabledCipherSuites,
                                    List<String> applicationProtocols);
